requirements.txt
flask==2.3.0
flask-cors==4.0.0
pymysql==1.1.0
PyJWT==2.8.0
